requirements.txt
# packages in environment at /home/duy/anaconda3/envs/bung_py3.8:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
_openmp_mutex             5.1                       1_gnu  
anyio                     4.3.0                    pypi_0    pypi
blas                      1.0                         mkl  
brotli-python             1.0.9            py38h6a678d5_7  
bzip2                     1.0.8                h7b6447c_0  
ca-certificates           2023.08.22           h06a4308_0  
certifi                   2023.7.22        py38h06a4308_0  
cffi                      1.15.1           py38h5eee18b_3  
charset-normalizer        2.0.4              pyhd3eb1b0_0  
coloredlogs               15.0.1                   pypi_0    pypi
cpuonly                   2.0                           0    pytorch
cryptography              41.0.3           py38hdda0065_0  
ffmpeg                    4.3                  hf484d3e_0    pytorch
filelock                  3.9.0            py38h06a4308_0  
filterpy                  1.4.5                    pypi_0    pypi
flatbuffers               24.3.7                   pypi_0    pypi
fonttools                 4.49.0                   pypi_0    pypi
freetype                  2.12.1               h4a9f257_0  
giflib                    5.2.1                h5eee18b_3  
gmp                       6.2.1                h295c915_3  
gmpy2                     2.1.2            py38heeb90bb_0  
gnutls                    3.6.15               he1e5248_0  
h11                       0.14.0                   pypi_0    pypi
httpcore                  1.0.4                    pypi_0    pypi
httpx                     0.27.0                   pypi_0    pypi
humanfriendly             10.0                     pypi_0    pypi
idna                      3.4              py38h06a4308_0  
imagehash                 4.3.1                    pypi_0    pypi
imageio                   2.34.0                   pypi_0    pypi
intel-openmp              2023.1.0         hdb19cb5_46306  
jinja2                    3.1.2            py38h06a4308_0  
jpeg                      9e                   h5eee18b_1  
lame                      3.100                h7b6447c_0  
lcms2                     2.12                 h3be6417_0  
ld_impl_linux-64          2.38                 h1181459_1  
lerc                      3.0                  h295c915_0  
libdeflate                1.17                 h5eee18b_1  
libffi                    3.4.4                h6a678d5_0  
libgcc-ng                 11.2.0               h1234567_1  
libgomp                   11.2.0               h1234567_1  
libiconv                  1.16                 h7f8727e_2  
libidn2                   2.3.4                h5eee18b_0  
libjpeg-turbo             2.0.0                h9bf148f_0    pytorch
libpng                    1.6.39               h5eee18b_0  
libstdcxx-ng              11.2.0               h1234567_1  
libtasn1                  4.19.0               h5eee18b_0  
libtiff                   4.5.1                h6a678d5_0  
libunistring              0.9.10               h27cfd23_0  
libwebp                   1.3.2                h11a3e52_0  
libwebp-base              1.3.2                h5eee18b_0  
llvm-openmp               14.0.6               h9e868ea_0  
lmdb                      1.4.1                    pypi_0    pypi
lz4-c                     1.9.4                h6a678d5_0  
markdown-it-py            3.0.0                    pypi_0    pypi
markupsafe                2.1.1            py38h7f8727e_0  
matplotlib                3.7.5                    pypi_0    pypi
mdurl                     0.1.2                    pypi_0    pypi
mkl                       2023.1.0         h213fc3f_46344  
mkl-service               2.4.0            py38h5eee18b_1  
mkl_fft                   1.3.8            py38h5eee18b_0  
mkl_random                1.2.4            py38hdb19cb5_0  
mpc                       1.1.0                h10f8cd9_1  
mpfr                      4.0.2                hb69a4c5_1  
mpmath                    1.3.0            py38h06a4308_0  
ncurses                   6.4                  h6a678d5_0  
nettle                    3.7.3                hbbd107a_1  
networkx                  3.1              py38h06a4308_0  
numpy                     1.24.3           py38hf6e8229_1  
numpy-base                1.24.3           py38h060ed82_1  
onnx                      1.15.0                   pypi_0    pypi
onnxruntime               1.17.1                   pypi_0    pypi
onnxsim                   0.4.36                   pypi_0    pypi
opencv-python             4.9.0.80                 pypi_0    pypi
openh264                  2.1.1                h4ff587b_0  
openjpeg                  2.4.0                h3ad879b_0  
openssl                   3.0.12               h7f8727e_0  
pexpect                   4.8.0              pyhd3eb1b0_3  
pillow                    10.0.1           py38ha6cbd5a_0  
pip                       23.3             py38h06a4308_0  
protobuf                  5.26.0                   pypi_0    pypi
ptyprocess                0.7.0              pyhd3eb1b0_2  
pycparser                 2.21               pyhd3eb1b0_0  
pyopenssl                 23.2.0           py38h06a4308_0  
pysocks                   1.7.1            py38h06a4308_0  
python                    3.8.18               h955ad1f_0  
pytorch                   2.1.0               py3.8_cpu_0    pytorch
pytorch-mutex             1.0                         cpu    pytorch
pyyaml                    6.0.1            py38h5eee18b_0  
readline                  8.2                  h5eee18b_0  
requests                  2.31.0           py38h06a4308_0  
rich                      13.7.1                   pypi_0    pypi
scikit-image              0.21.0                   pypi_0    pypi
seaborn                   0.13.0                   pypi_0    pypi
setuptools                68.0.0           py38h06a4308_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.41.2               h5eee18b_0  
sympy                     1.11.1           py38h06a4308_0  
tbb                       2021.8.0             hdb19cb5_0  
thop                      0.1.1-2209072238          pypi_0    pypi
tk                        8.6.12               h1ccaba5_0  
torchaudio                2.1.0                  py38_cpu    pytorch
torchvision               0.16.0                 py38_cpu    pytorch
typing_extensions         4.7.1            py38h06a4308_0  
ultralytics               8.0.220                  pypi_0    pypi
urllib3                   1.26.18          py38h06a4308_0  
wheel                     0.41.2           py38h06a4308_0  
xz                        5.4.2                h5eee18b_0  
yaml                      0.2.5                h7b6447c_0  
zlib                      1.2.13               h5eee18b_0  
zstd                      1.5.5                hc292b87_0  
